+++ src/hooks/index.ts
@@ -1,6 +1,7 @@
 import useFixtures from "./useFixtures";
+import usePlayers from "./usePlayers";
 import useSeasons from "./useSeasons";
 import useTables from "./useTables";
 import useTeams from "./useTeams";
 
-export { useFixtures, useSeasons, useTables, useTeams };
+export { useFixtures, usePlayers, useSeasons, useTables, useTeams };